requirements.txt
numpy
pandas
scipy
scikit-learn
yfinance
fredapi
requests
urllib3
pyyaml
streamlit
plotly
altair
matplotlib
python-dateutil
cachetools